add automation cache creation
When new modules are committed, automation needs to rebuild the cache.

The scripts provided here will utilize the public docker container to
rebuild the cache json file. One drawback to this approach is that the
container must be up to date in dockerhub before the rebuild occurs.

Ideas on a better way to do this without needing to build the docker
container in each automation run would be appreciated.